requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
inflection==0.5.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
